requirements.txt
fastapi==0.111.0
uvicorn==0.26.1
requests==2.32.4
python-dotenv==1.0.1
pydantic==2.11.7
